replaced old main with new one
